+++ db/db_range_del_test.cc
@@ -504,7 +504,7 @@ TEST_F(DBRangeDelTest, ValidUniversalSubcompactionBoundaries) {
       1 /* input_level */, 2 /* output_level */, CompactRangeOptions(),
       nullptr /* begin */, nullptr /* end */, true /* exclusive */,
       true /* disallow_trivial_move */,
-      std::numeric_limits<uint64_t>::max() /* max_file_num_to_ignore */,
+      (size_t)std::numeric_limits<uint64_t>::max() /* max_file_num_to_ignore */,
       "" /*trim_ts*/));
 }
 #endif  // ROCKSDB_LITE
