+++ table/block_based/block_based_table_reader_sync_and_async.h
@@ -355,7 +355,7 @@ DEFINE_SYNC_AND_ASYNC(void, BlockBasedTable::MultiGet)
       iiter_unique_ptr.reset(iiter);
     }
 
-    uint64_t prev_offset = std::numeric_limits<uint64_t>::max();
+    uint64_t prev_offset = (size_t)std::numeric_limits<uint64_t>::max();
     autovector<BlockHandle, MultiGetContext::MAX_BATCH_SIZE> block_handles;
     autovector<CachableEntry<Block>, MultiGetContext::MAX_BATCH_SIZE> results;
     autovector<Status, MultiGetContext::MAX_BATCH_SIZE> statuses;
